feat(catalog): implement series detail page and support float ratings
- Add `SeriesShowLive` LiveView to handle series details and interactions.
- Migrate `rating` and `rating_sum` from integer to float to support
  half-star increments (0.5).
- Implement `rating_distribution` logic to bucket half-star ratings
  into a stable 5-row UI.
- Enhance `Catalog.Query` with search and sorting for chapters and comments.
- Update admin dashboard to support improved series and tag management.
- Update database migrations and tests to accommodate float rating types.
- Refactor S3 storage configuration and improve runtime validation.

+++ lib/crysa/catalog/series.ex
@@ -46,7 +46,7 @@ defmodule Crysa.Catalog.Series do
     field :bookmark_count, :integer, default: 0
     field :view_count, :integer, default: 0
     field :rating_count, :integer, default: 0
-    field :rating_sum, :integer, default: 0
+    field :rating_sum, :float, default: 0.0
     field :next_check_at, :utc_datetime_usec
     field :last_checked_at, :utc_datetime_usec
